+++ feat/wikiDustLoop/src/commonTest/kotlin/io/github/sophon/wikidustloop/data/MoveMapperTest.kt
@@ -3,12 +3,14 @@ package io.github.sophon.wikidustloop.data
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import io.github.sophon.core.feature.Game
+import io.github.sophon.core.wiki.usecase.DownloadMoveListUseCase
 import kotlin.test.Test
 
 class MoveMapperTest {
     val gb = Game.GBVSR.id
     val gg = Game.GGST.id
     val bb = Game.BBCF.id
+    val emptyCharData = DownloadMoveListUseCase.CharacterData("", null)
 
     //region Aliases
     @Test
@@ -131,42 +133,8 @@ class MoveMapperTest {
     @Test
     fun `formWikiUrl handles spaces in inputs`() {
         // given
-        val string = "5D Bat"
         val expected = "https://www.dustloop.com/w/BBCF/Platinum_the_Trinity#Magical_Bat"
-        val dto = MoveDto(
-            chara = "Platinum the Trinity",
-            name = "Magical Bat",
-            input = "5D Bat",
-            damage = "900",
-            guard = "Mid",
-            startup = "11",
-            active = "4",
-            recovery = "26",
-            onBlock = "-16",
-            onODR = "-14",
-            attribute = "B",
-            invuln = "1~11 All",
-            cancel = "(S)R",
-            p1 = "60",
-            p2 = "75",
-            starter = "Very Short",
-            level = "2",
-            blockstun = "13",
-            groundHit = "Launch",
-            airHit = "30 + WStick 25",
-            groundCH = "Launch",
-            airCH = "42 + WBounce 40 + WStick 25",
-            blockstop = "22",
-            hitstop = "+0",
-            CHstop = "+1",
-            cancelTiming = null,
-            images = "BBCS_Platinum_homerun5D.png",
-            caption = "Also good in combos",
-            hitboxes = "BBCF Platinum homerun5D hitbox 1.png;BBCF Platinum homerun5D hitbox 2.png",
-            hitboxCaption = "Ground, frame 11\\Ground, frames 12-14",
-            type = "drive",
-            notes = "Counter Hit state for entire move; Reversal"
-        )
+        val dto = MoveSource.magicalBat
 
         // when
         val result = formMoveWikiUrl(bb, dto)
@@ -174,4 +142,88 @@ class MoveMapperTest {
         //then
         assertThat(result).isEqualTo(expected)
     }
+    
+    @Test
+    fun `toDomain handles close input`() {
+        // given
+        val clC = MoveSource.closeSlash
+        val expected = "c.s"
+        
+        // when
+        val result = clC.toDomain(
+            gameId = gg,
+            characterData = emptyCharData,
+            imageUrlMap = emptyMap(),
+        )
+
+        //then
+        assertThat(result.input).isEqualTo(expected)
+    }
+}
+
+private object MoveSource {
+    val closeSlash = MoveDto(
+        chara = "Sol Badguy",
+        name = null,
+        input = "c.S",
+        damage = "44",
+        guard = "All",
+        startup = "7",
+        active = "6",
+        recovery = "10",
+        onBlock = "+3",
+        onHit = "+13",
+        level = "4",
+        counter = "Mid",
+        images = "GGST Sol Badguy cS.png",
+        hitboxes = "GGST Sol cS Hitbox.png",
+        notes = "Input Proximity Range: 240;Hitstop on ground hit: 16F; Floating crumple on ground hit: Total 28F (airborne hitstun 1-11F, standing hitstun 12-18F, can block 19~28F)",
+        type = "normal",
+        riscGain = "1700",
+        riscLoss = "1000",
+        wallDamage = "300",
+        inputTension = null,
+        chipRatio = null,
+        OTGType = "Up",
+        prorate = "100%",
+        invuln = null,
+        cancel = "SJDRP",
+        caption = "Kills your opponent on block or hit",
+        hitboxCaption = "The sword also has a hitbox for some reason",
+    )
+    
+    val magicalBat = MoveDto(
+        chara = "Platinum the Trinity",
+        name = "Magical Bat",
+        input = "5D Bat",
+        damage = "900",
+        guard = "Mid",
+        startup = "11",
+        active = "4",
+        recovery = "26",
+        onBlock = "-16",
+        onODR = "-14",
+        attribute = "B",
+        invuln = "1~11 All",
+        cancel = "(S)R",
+        p1 = "60",
+        p2 = "75",
+        starter = "Very Short",
+        level = "2",
+        blockstun = "13",
+        groundHit = "Launch",
+        airHit = "30 + WStick 25",
+        groundCH = "Launch",
+        airCH = "42 + WBounce 40 + WStick 25",
+        blockstop = "22",
+        hitstop = "+0",
+        CHstop = "+1",
+        cancelTiming = null,
+        images = "BBCS_Platinum_homerun5D.png",
+        caption = "Also good in combos",
+        hitboxes = "BBCF Platinum homerun5D hitbox 1.png;BBCF Platinum homerun5D hitbox 2.png",
+        hitboxCaption = "Ground, frame 11\\Ground, frames 12-14",
+        type = "drive",
+        notes = "Counter Hit state for entire move; Reversal"
+    )
 }