+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # BOT CHANGELOG
 
+## [v11.0.0] - 2026-03-25
+- `stats` - show the total bot usage count
+- `SuperCombo`
+  - now supports motion inputs for SF
+  - fixed Blanka's Rolling Cannon embed
+- `DreamCancel` 
+  - fixed issue multi-button ie
+    - `236236b/d` now can be queried with both `236236b` or `236236d`
+- guard Discord command registration and cleanup
+- guard Kord embed creation and button interactions
+- fixed `cl` input error
+
 ## [v10.5.0] - 2026-03-14
 - `fd` - fixed not being able to search by move name or alias
 - remove redundant EWGF command